+++ packages/flutter_riverpod/lib/src/consumer.dart
@@ -17,11 +17,49 @@ abstract class WidgetRef {
 
   /// Listen to a provider and call `listener` whenever its value changes.
   ///
+  /// It is safe to call [listen] within the `build` method of a widget:
+  ///
+  /// ```dart
+  /// Consumer(
+  ///   builder: (context, ref, child) {
+  ///     ref.listen<int>(counterProvider, (prev, next) {
+  ///       print('counter changed $next');
+  ///     });
+  ///   },
+  /// )
+  /// ```
+  ///
+  /// When used inside `build`, listeners will automatically be removed
+  /// if a widget rebuilds and stops listening to a provider.
+  ///
+  /// [listen] should not be used outside of the `build` method of a widget.
+  /// It cannot be used within widget life-cycles such as initState/dispose or
+  /// event handles such as "on tap".
+  ///
   /// This is useful for showing modals or other imperative logic.
   void listen<T>(
     ProviderListenable<T> provider,
     void Function(T? previous, T next) listener, {
     void Function(Object error, StackTrace stackTrace)? onError,
+  });
+
+  /// Listen to a provider and call `listener` whenever its value changes.
+  ///
+  /// As opposed to [listen], [listenOnce] is not safe to use within the `build`
+  /// method of a widget.
+  /// [listenOnce] is meant to typically be used inside [State.initState].
+  ///
+  /// [listenOnce] returns a [ProviderSubscription] which can be used to stop
+  /// listening to the provider, or to read the current value exposed by
+  /// the provider.
+  ///
+  /// It is not necessary to call [ProviderSubscription.close]. The
+  /// subsription will automatically be closed when the widget is disposed.
+  ProviderSubscription<T> listenOnce<T>(
+    ProviderListenable<T> provider,
+    void Function(T? previous, T next) listener, {
+    void Function(Object error, StackTrace stackTrace)? onError,
+    bool fireImmediately,
   });
 
   /// Reads a provider without listening to it.
@@ -419,6 +457,7 @@ class ConsumerStatefulElement extends StatefulElement implements WidgetRef {
   var _dependencies = <ProviderListenable, ProviderSubscription>{};
   Map<ProviderListenable, ProviderSubscription>? _oldDependencies;
   final _listeners = <ProviderSubscription>[];
+  List<_ListenOnce>? _onceListeners;
 
   @override
   void didChangeDependencies() {
@@ -476,6 +515,14 @@ class ConsumerStatefulElement extends StatefulElement implements WidgetRef {
     for (var i = 0; i < _listeners.length; i++) {
       _listeners[i].close();
     }
+    final onceListeners = _onceListeners;
+    if (onceListeners != null) {
+      for (var i = 0; i < onceListeners.length; i++) {
+        onceListeners[i].close();
+      }
+      _onceListeners = null;
+    }
+
     super.unmount();
   }
 
@@ -511,4 +558,43 @@ class ConsumerStatefulElement extends StatefulElement implements WidgetRef {
   void invalidate(ProviderBase<Object?> provider) {
     _container.invalidate(provider);
   }
+
+  @override
+  ProviderSubscription<T> listenOnce<T>(
+    ProviderListenable<T> provider,
+    void Function(T? previous, T next) listener, {
+    void Function(Object error, StackTrace stackTrace)? onError,
+    bool fireImmediately = false,
+  }) {
+    final listeners = _onceListeners ??= [];
+
+    final sub = _ListenOnce(
+      ProviderScope.containerOf(this, listen: false).listen(
+        provider,
+        listener,
+        onError: onError,
+        fireImmediately: fireImmediately,
+      ),
+      this,
+    );
+    listeners.add(sub);
+
+    return sub;
+  }
+}
+
+class _ListenOnce<T> implements ProviderSubscription<T> {
+  _ListenOnce(this._subscription, this._element);
+
+  final ProviderSubscription<T> _subscription;
+  final ConsumerStatefulElement _element;
+
+  @override
+  void close() {
+    _subscription.close();
+    _element._onceListeners?.remove(this);
+  }
+
+  @override
+  T read() => _subscription.read();
 }